+++ BrickBreak/src/Level.h
@@ -10,7 +10,7 @@ class Level
 public:
     std::vector<Object> objects;
 
-    Level(const char* file, unsigned int levelWidth, unsigned int levelHeight);
+    Level(const char* file, float levelWidth, float levelHeight);
     void load();
     void unload();
 
@@ -20,9 +20,9 @@ public:
     inline bool isLoaded() const { return m_IsLoaded; }
 private:
     const char* m_File;
-    unsigned int m_Width;
-    unsigned int m_Height;
+    float m_Width;
+    float m_Height;
     bool m_IsLoaded = false;
 
-    void init(std::vector<std::vector<unsigned int>>& tilesData, unsigned int lvlWidth, unsigned int lvlHeight);
+    void init(std::vector<std::vector<unsigned int>>& tilesData, float lvlWidth, float lvlHeight);
 };
